requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.2
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.231
transformers>=4.36.0
faiss-cpu>=1.7.4
opencv-python>=4.8.1.78
requests>=2.31.0
werkzeug==3.0.1
tqdm>=4.66.0
pandas>=2.0.0

